Engine Folder Openning + Compiler warning fixes (#155)
* Engine Folder Openning + Compiler warning fixes

* Drag Drop multiple files.

* able to drop on tree.

* removed redundant null check.

* copy, cut, paste with Ctrl+C, Ctrl+V

* bug fix.

* Left click copy paste works.

* no need g_current entry.

* esc discard selection, shift instead of ctrl.

* bug, feedback fixes.

* more comments.

* last feedback fixes. ctrl deselect, seperate FolderView.cpp

* cutting, copying highlight colors. and fix right click behaviour

---------
