+++ Editor/PluginWindow.cpp
@@ -33,7 +33,7 @@ namespace ToolKit
       m_screenResolutions.push_back(Vec2(500.0f, 500.0f));
     }
 
-    void PluginWindow::RemoveResolutionName(int index)
+    void PluginWindow::RemoveResolutionName(size_t index)
     {
       bool canRemove = index > 0 || index < m_screenResolutions.size();
